requirements.txt
psycopg2
pika
cbpro
python-dateutil